\name{alleleFinder}
\alias{alleleFinder}
\title{
Scores the diploid genotypes into an output table
}
\description{
Estimates the diploid genotypes based on read size (primer to primer) in bp. The results are written to two files: one with two alleles
 per column and the other with the two alleles collapsed in a single column. The files are written to a folder named GenotypesOut
created by the function. 
}
\usage{
alleleFinder(sampleDir="./samplesF",locusInfo="primers",calibrationFiles="CalibrationSamples.txt")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{sampleDir}{
The directory where the filtered reads are stored. This directory should only contain the read files and 
nothing else. By default this is a directory named samplesF written in the session working directory, 
as produced by \code{\link{read.quality.filter}}
}
\item{locusInfo}{
A tab delimited text file containing the information for the microsatellite loci used. Should contain four columns with the following header names:\cr
\cr
Locus: The loci names\cr
\cr
Fprimer: The sequence of the FWD primer in upper case\cr
\cr
Rprimer: The sequence of the RVS primer in upper case\cr
\cr
Motif: The sequence for the microsatellite motif (e.g., CTTG for tetranucleotide). Note that for some loci you might need to use the reverse
 complementar sequence of what your reference value is. Thus, inspect your reads first or if you are not getting data for some loci try this out.
 For example, for a motif CAGG, you might need to use CCTG.\cr
\cr
nmotif: Reads are retained if they have at least these many repetitions of the repeat motif.
}
\item{calibrationFiles}{
The file described below is the same as used for function  \code{\link{GenotypeModel}}, earlier in the previous pipeline step. Here the only information used is the last row containing the sequencing coverage threshold for each marker (number of reads of similar sequence length) that will applied to filter alleles scored. Inspect the plots produced by function  \code{\link{reads.plots}} carefully to select values for coverage filtering. Essentialy use a value below which you don't trust reads to be true alleles.  
A tab delimited file with as many columns as loci amplified in the samples. The order of loci is required to match the order in the locusInfo file. The header names must start with the locus name, written exactly has in the file passed to locusInfo. The first row contains the range of alleles for each locus, the format should be 100-250, for an example of an allelic range between 100 and 250 bp. Users should inspect the plots created by \code{\link{reads.plots}} to determine the ranges. Be conservative and add a buffer of about 10 bp on each side but know that the larger the range the slower the function will be. The following rows contain samplenames (one per row) for a homozygote samples that will be used to generate the expected relative frequencies under different combinations of alleles.
Users can select these homozygote representative samples by inspecting the plots saved in the Plots directory (generated by \code{\link{reads.plots}}). Use the exact sample names as the file names stored in directory samplesF. For convenience, sample names are  written to the plots and can be copied from there when selecting homozygote samples. 
}
}
\details{
Uses the tables created by \code{\link{GenotypeModel}}, containing expected relative frequencies of read cover for different combinations (2 out of N) of alleles,
to contrast with observed data in each sample and find the diploid genotypes minimizing a sum of squares. For each observed sample, the row table that minimizes 
the sum of squares: sum((ORF(a)-ERF(a))^2), with ORF and ERF the Observed and Expected relative frequencies at each read size, is the estimated genotype.

}
\value{
A data frame with a concatenated genotype (the two alleles in the same cell) is returned. This table is also written to file, as well as a version with 
the two alleles in one locus in different columns (two columns per locus). The output files are written to a folder named GenotypesOut that is created by 
the function in the R session working directory.
}
\references{
Alberto F (in prep)...
}
\author{
F. Alberto
}
\seealso{
\code{\link{GenotypeModel}}, \code{\link{AlleleCum}}
}
\keyword{ IO }
