\name{GenotypeModel}
\alias{GenotypeModel}
\title{
Creates tables with expected relative frequencies of read sizes
}
\description{
Creates tables with expected relative frequencies of read sizes for each possible diploid genotype within an allelic range set by the user.
}
\usage{
GenotypeModel(sampleDir="./samplesF",locusInfo="primers",drop.out.b=0.015,
              calibrationFiles="CalibrationSamples.txt")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{sampleDir}{
The directory where the filtered reads are stored. This directory should only contain the read files and 
nothing else. By default this is a directory named samplesF written in the session working directory, 
as produced by \code{\link{read.quality.filter}}
}
\item{locusInfo}{
A tab delimited text file containing the information for the microsatellite loci used. Should contain four columns with the following header names:\cr
\cr
Locus: The loci names\cr
\cr
Fprimer: The sequence of the FWD primer in upper case\cr
\cr
Rprimer: The sequence of the RVS primer in upper case\cr
\cr
Motif: The sequence for the microsatellite motif (e.g., CTTG for tetranucleotide). Note that for some loci you might need to use the reverse
 complementar sequence of what your reference value is. Thus, inspect your reads first or if you are not getting data for some loci try this out.
 For example, for a motif CAGG, you might need to use CCTG.\cr
\cr
nmotif: Reads are retained if they have at least these many repetitions of the repeat motif.
}
\item{drop.out.b}{
The proportion of decline in coverage per base pair distance between small and large allele on a heterozygous loci. It defauts to 0.015. It could be estimated from the data from each locus, in a future version, if the user supplies a list of heterozygous samples (upon inspecting the plots produced).
}
\item{calibrationFiles}{
A tab delimited file with as many columns as loci amplified in the samples. The order of loci is required to match the order in the locusInfo file. The header names must start with the locus name, written exactly has in the file passed to locusInfo. The first row contains the range of alleles for each locus, the format should be 100-250, for an example of an allelic range between 100 and 250 bp. Users should inspect the plots created by \code{\link{reads.plots}} to determine the ranges. Be conservative and add a buffer of about 10 bp on each side but know that the larger the range the slower the function will be. The following rows contain samplenames (one per row) for a homozygote samples that will be used to generate the expected relative frequencies under different combinations of alleles.
Users can select these homozygote representative samples by inspecting the plots saved in the Plots directory (generated by \code{\link{reads.plots}}). Use the exact sample names as the file names stored in directory samplesF. For convenience, sample names are  written to the plots and can be copied from there when selecting homozygote samples. 
The final row contains the coverage threshold (number of reads of similar sequence length) that will applied to filter alleles scored. This function will not use this information but to keep the format of the calibration file similar across this and other functions the information is required here.
}
}
\details{
Using a set of representative homozygote samples and an allelic range provided by the user, the function produces a large table with as many rows as combinations
of diploid genotypes, i.e., all combinations of 2 in N, N being defined by seq(lower.range, upper.range, 1). To these combination of 2 in N the homozygote cases,
 for all alleles in the defined range, are added as additional possible genotypes. Then a pattern of relative frequencies from a single allele, generated from
  the analysis of reads in the homozygote reference samples, is shifted along the allelic range to match the bp position of each different combination of alleles.
  This procedure generates all the possible overlaps of read relative frequencies for the heterozygote cases. For heterozygotes the relative frequencies are then 
  divided by two so that the row sum adds to 1. The table rows (one for each possible genotype) are then compared against the observed data for each sample by
  function \code{\link{alleleFinder}} to find the row table that minimizes a sum of squares relative to the observed relative frequencies for each read in each sample:
   sum((ORF(a)-ERF(a))^2), with ORF and ERF the Observed and Expected relative frequencies at each read size a. The row with the minimum sum is the estimated genotype.
}
\value{
One model table for each locus  is written in a directory, named ModelTables, created by the function in the working directory. 
}
\references{
Alberto F (in prep)...
}
\author{
F. Alberto
}
\seealso{
\code{\link{alleleFinder}}
}
\keyword{ IO }
