\name{M.allele.Finder}
\alias{M.allele.Finder}
\title{
Identifies multiple alleles (possibly more than two) for each loci in each sample.
}
\description{
The function searches samples for multiple alleles of the same loci. Examples of such samples come from forensic analysis, where DNA from multiple individuals might be present, or eDNA samples containing alleles from several individuals. Only alleles present in the reference population can be detected. Plots with the scored alleles are written to a new directory created in the working environment *M.plots*.
}
\usage{
M.allele.Finder(sampleDir="./samplesF.MA", locusInfo="primers",
               calibrationFiles="CalibrationSamples.txt",
               ReadSizesRepeatNumbers="./ReadSizesRepeatNumbers/",
               RefGenotypes="./GenotypesOut/Genotypes allele table.txt")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{sampleDir}{
The directory where the filtered reads are stored. This directory should only contain the read files and 
nothing else. By default this is a directory named samplesF written in the session working directory, 
as produced by \code{\link{read.quality.filter}}
}
\item{locusInfo}{
A tab delimited text file containing the information for the microsatellite loci used. Should contain four columns with the following header names:\cr
\cr
Locus: The loci names\cr
\cr
Fprimer: The sequence of the FWD primer in upper case\cr
\cr
Rprimer: The sequence of the RVS primer in upper case\cr
\cr
Motif: The sequence for the microsatellite motif (e.g., CTTG for tetranucleotide). Note that for some loci you might need to use the reverse
 complementar sequence of what your reference value is. Thus, inspect your reads first or if you are not getting data for some loci try this out.
 For example, for a motif CAGG, you might need to use CCTG.\cr
\cr
nmotif: Reads are retained if they have at least these many repetitions of the repeat motif.
}
\item{calibrationFiles}{
A tab delimited file with as many columns as loci amplified in the samples. The header names must start with the locus name, written exactly has in 
the file passed to locusInfo. The first row contains the range of alleles for each locus, the format should be 100-250, for an example of an allelic 
range between 100 and 250 bp. Users should inspect the plots created by \code{\link{reads.plots}} to determine the ranges. Be conservative and 
add a buffer of about 10 bp on each side but know that the larger the range the slower the function will be. The following rows contain sample
names (one per row) for a homozygote samples that will be used to generate the expected relative frequencies under different combinations of alleles.
Users can select these homozygote representative samples by inspecting the plots saved in the Plots directory (generated by \code{\link{reads.plots}}).
 Use the exact sample names as the file names stored in directory samplesF. For convenience, sample names are  written to the plots 
and can be copied from there when selecting homozygote samples. I used six samples per locus in my test (thus my calibration table had a total of seven rows).
}
\item{ReadSizesRepeatNumbers}{
The path to the folder with the same name created by function \code{\link{read.size.freqs}}.
}
 \item{RefGenotypes}{
A string with the path for the reference file genotype file in the 2 columns per locus format. The format is the same as produced by function
\code{\link{alleleFinder}} and thus the default path is the folder where that file is written by this function "./GenotypesOut/Genotypes allele table.txt". Only alleles present in the reference population can be found.
}
}
\details{
The function searches for the presence of alleles, from those present in the reference population, in samples that may have up to n alleles. Examples of such samples come from forensic analysis where DNA from multiple multiple individuals might be present or eDNA samples containing alleles from several individuals.
}
\value{
The function returns to the working environment a list with the alleles found for each locus and individual. The function also prints to the working directory a pdf file for each sample containing plots of read (allele) distribution, with separate plots per page showing the different loci, and red vertical lines indicating the alleles scored.
}
\references{
Alberto F (in prep)...
}
\author{
F. Alberto
}
\seealso{
\code{\link{GenotypeModel}}, \code{\link{alleleFinder}}
}
\keyword{ IO }
